Change what goal means

+++ internal/common/campaigns.go
@@ -32,7 +32,7 @@ type Campaign struct {
 	Status   bool  `json:"status"`
 	Approved int32 `json:"approved"` // Set to ts when admin receives all perks (or there are no perks)
 
-	Goal int `json:"infGoal"` // How many influencers does the client want ideally as part of this deal? Ignored if perks present
+	Goal float64 `json:"infGoal"` // Price per influencer goal
 
 	// Social Media Post/User Requirements
 	Tags    []string         `json:"tags,omitempty"`
